--- package.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>sub8_ros_tools</name>
-  <version>1.0.0</version>
-  <description>The sub8_ros_tools package</description>
-  <maintainer email="jpanikulam@ufl.edu">Jacob Panikulam</maintainer>
-  <license>MIT</license>
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>rostest</build_depend>
-  <run_depend>rospy</run_depend>
-  <export>
-  </export>
-</package>
