--- readme.md
@@ -1,10 +1,0 @@
-sub8_ros_tools
-==============
-
-Miscellaneous ROS tools that are useful for various tasks.
-
-Examples:
-
-# TODO
-* The items that directly interact with ros topics or the parameter server do not yet have unittests, they require rostest.
-
