--- sub8_misc_tools/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-from download import download_and_unzip
-from download import download
-
-import text_effects
