--- sub8_tools/__init__.py
@@ -1,2 +1,0 @@
-from sub8_ros_tools import *
-from sub8_misc_tools import *
